pyproject.toml
[tool.poetry]
name = "cv-genrator"
version = "0.1.0"
description = ""
authors = ["Gabriel <115580274+gabszs@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<=3.13"
fastapi = {extras = ["all"], version = "^0.110.0"}
pydantic-settings = "^2.2.1"
icecream = "^2.1.3"
sqlalchemy = "^2.0.28"
redis = "^5.0.3"
dependency-injector = "^4.41.0"
python-jose = "^3.3.0"
pyjwt = "^2.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
psycopg = {extras = ["binary"], version = "^3.1.18"}
asyncpg = "^0.29.0"
crewai = "^0.22.5"
alembic = "^1.13.1"

[tool.poetry.group.dev.dependencies]
pytest-cov = "^4.1.0"
taskipy = "^1.12.2"
ruff = "^0.1.8"
icecream = "^2.1.3"
pytest-asyncio = "^0.23.2"
factory-boy = "^3.3.0"
freezegun = "^1.4.0"
faker = "^24.1.0"
ipykernel = "^6.29.3"
pytest = "^8.1.1"
pytest-postgresql = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
run = 'uvicorn app.main:app --reload'
lint = 'ruff .'
pre_test = 'task lint'
test = 'pytest -s -x --capture=no --cov=app -vv'
verbose_test = 'pytest --verbose --show-capture=all --exitfirst --cov=app --cov-report=term-missing -vv'
commit_hook = "pre-commit run --all-files"
post_verbose_test = 'coverage html'
post_test = 'coverage html'

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 120
indent-width = 4

# Assume Python 3.8
target-version = "py38"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"
